packages/graphrag-vectors/pyproject.toml
[project]
name = "graphrag-vectors"
version = "3.0.8"
description = "GraphRAG vector store package."
authors = [
    {name = "Alonso Guevara Fernández", email = "alonsog@microsoft.com"},
    {name = "Andrés Morales Esquivel", email = "andresmor@microsoft.com"},
    {name = "Chris Trevino", email = "chtrevin@microsoft.com"},
    {name = "David Tittsworth", email = "datittsw@microsoft.com"},
    {name = "Dayenne de Souza", email = "ddesouza@microsoft.com"},
    {name = "Derek Worthen", email = "deworthe@microsoft.com"},
    {name = "Gaudy Blanco Meneses", email = "gaudyb@microsoft.com"},
    {name = "Ha Trinh", email = "trinhha@microsoft.com"},
    {name = "Jonathan Larson", email = "jolarso@microsoft.com"},
    {name = "Josh Bradley", email = "joshbradley@microsoft.com"},
    {name = "Kate Lytvynets", email = "kalytv@microsoft.com"},
    {name = "Kenny Zhang", email = "zhangken@microsoft.com"},
    {name = "Mónica Carvajal"},
    {name = "Nathan Evans", email = "naevans@microsoft.com"},
    {name = "Rodrigo Racanicci", email = "rracanicci@microsoft.com"},
    {name = "Sarah Smith", email = "smithsarah@microsoft.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "azure-core~=1.32",
    "azure-cosmos~=4.9",
    "azure-identity~=1.25",
    "azure-search-documents~=11.6",
    "graphrag-common==3.0.8",
    "lancedb~=0.24.1",
    "numpy~=2.1",
    "pyarrow~=22.0",
    "pydantic~=2.10",
    # GCP optional extras (Spanner + Vertex AI vector store backends)
    "google-cloud-spanner>=3.40,<4",
    "google-cloud-aiplatform>=1.60,<2",
    "graphrag-storage==3.0.8",
]

[project.urls]
Source = "https://github.com/microsoft/graphrag"

[build-system]
requires = ["hatchling>=1.27.0,<2.0.0"]
build-backend = "hatchling.build"
